requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
googletrans==3.1.0a0
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.5.0
playsound==1.3.0
pytesseract==0.3.8
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
urllib3==1.26.8
Flask-Dropzone==1.6.0